Split futures private actions into separate modules
