+++ lib/ex_okex/futures/private.ex
@@ -5,14 +5,11 @@ defmodule ExOkex.Futures.Private do
   [API docs](https://www.okex.com/docs/en/#futures-README)
   """
 
-  import ExOkex.Api.Private
   alias ExOkex.Futures.Private
 
   @type params :: map
   @type config :: ExOkex.Config.t()
   @type response :: ExOkex.Api.response()
-
-  @prefix "/api/futures/v3"
 
   @doc """
   Place a new order.
@@ -32,9 +29,7 @@ defmodule ExOkex.Futures.Private do
   {:ok, %{"order_info" => [%{"error_code" => 0, "error_message" => "", "order_id" => "2653481276189696"}], "result" => true}}
   """
   @spec create_order(params, config | nil) :: response
-  def create_order(params, config \\ nil) do
-    post("#{@prefix}/orders", params, config)
-  end
+  defdelegate create_order(params, config \\ nil), to: Private.CreateOrder
 
   @doc """
   Place multiple orders for specific trading pairs (up to 4 trading pairs, maximum 4 orders each)
@@ -52,15 +47,13 @@ defmodule ExOkex.Futures.Private do
       "leverage":"10" },
     ])
 
-    # TODO: Add response sample
-
   """
   @spec create_bulk_orders([params], config | nil) :: response
-  def create_bulk_orders(params, config \\ nil) do
-    post("#{@prefix}/orders", params, config)
-  end
+  defdelegate create_bulk_orders(params, config \\ nil), to: Private.CreateBulkOrders
 
-  defdelegate create_batch_orders(params, config \\ nil), to: __MODULE__, as: :create_bulk_orders
+  defdelegate create_batch_orders(params, config \\ nil),
+    to: Private.CreateBulkOrders,
+    as: :create_bulk_orders
 
   @doc """
   Cancelling an unfilled order.
@@ -69,14 +62,11 @@ defmodule ExOkex.Futures.Private do
 
   ## Example
 
-      iex> ExOkex.Futures.cancel_orders("BTC-USD-180309", [1600593327162368,1600593327162369])
+  iex> ExOkex.Futures.cancel_orders("BTC-USD-180309", [1600593327162368,1600593327162369])
 
-      # TODO: Add response
   """
-  def cancel_orders(instrument_id, order_ids \\ [], params \\ %{}, config \\ nil) do
-    new_params = Map.merge(params, %{order_ids: order_ids})
-    post("#{@prefix}/cancel_batch_orders/#{instrument_id}", new_params, config)
-  end
+  defdelegate cancel_orders(instrument_id, order_ids \\ [], params \\ %{}, config \\ nil),
+    to: Private.CancelOrders
 
   @doc """
   Get the futures account info of all token.
@@ -85,13 +75,10 @@ defmodule ExOkex.Futures.Private do
 
   ## Examples
 
-      iex(3)> ExOkex.Futures.list_accounts()
+  iex> ExOkex.Futures.list_accounts()
 
-      # TODO: Add Response
   """
-  def list_accounts(config \\ nil) do
-    get("#{@prefix}/accounts", %{}, config)
-  end
+  defdelegate list_accounts(config \\ nil), to: Private.ListAccounts
 
   @doc """
   Retrieve information on your positions of a single contract.
@@ -100,7 +87,7 @@ defmodule ExOkex.Futures.Private do
 
   ## Examples
 
-      iex(3)> ExOkex.Futures.Private.position("BTC-USD-190329")
+  iex> ExOkex.Futures.Private.position("BTC-USD-190329")
 
   """
   defdelegate position(instrument_id, config \\ nil), to: Private.Position
@@ -112,7 +99,7 @@ defmodule ExOkex.Futures.Private do
 
   ## Examples
 
-      iex(3)> ExOkex.Futures.Private.list_positions()
+  iex> ExOkex.Futures.Private.list_positions()
 
   """
   defdelegate list_positions(config \\ nil), to: Private.ListPositions
