+++ lib/ex_okex/futures/private/create_bulk_orders.ex
@@ -1,0 +1,10 @@
+defmodule ExOkex.Futures.Private.CreateBulkOrders do
+  import ExOkex.Api.Private
+
+  @prefix "/api/futures/v3"
+
+  def create_bulk_orders(params, config \\ nil) do
+    "#{@prefix}/orders"
+    |> post(params, config)
+  end
+end
