+++ lib/ex_okex/futures/private/create_order.ex
@@ -1,0 +1,10 @@
+defmodule ExOkex.Futures.Private.CreateOrder do
+  import ExOkex.Api.Private
+
+  @prefix "/api/futures/v3"
+
+  def create_order(params, config \\ nil) do
+    "#{@prefix}/orders"
+    |> post(params, config)
+  end
+end
